+++ HISTORY.md
@@ -6,12 +6,14 @@
     * fixed recursive calls in PIPE
     * the "length of list" block no longer appears twice in search results
     * prevent inserting items at non-integer / out-of-bounds indices
+    * save all items of a heterogeneously structured linked/arrayed list
 
 ### 2021-03-09
 * new dev version
 * objects: fixed #2797
 * fixed recursive calls in PIPE
 * lists: prevent usage of lists as dictionaries
+* store: fixed ##2798
 
 ## 6.7.0
 * **New Features:**